Updated module augmentation for code-block package.
